apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.1
celery[redis]==5.3.6
qdrant-client==1.7.3
sentence-transformers==2.3.1
cryptography==42.0.2
stripe==8.2.0
pyyaml==6.0.1
structlog==24.1.0
tenacity==8.2.3
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-httpx==0.30.0
